added Collection base class
Created the Collection base class as a starting point for the
implementation of Bloom tables.  Currently it passes a single test in
the simplest possible manner.  However the intent is to expand the
test suites to exercise the relational model of Collections to support
key columns, aggregates, projections, and the whole nine yards.

The Collection class will serve as a base class for the actual Bloom
table objects.