Add KeyConstraintError check to insert in Collection
When inserting a new tuple to the collection we check to see if there
is an existing tuple in the collection with for the computed key.  If
there is we check to see if the new and the old tuple disagree on the
value of the non-key columns.  If they do, we raise
KeyConstraintError; facts are immutable in collections.